update mkrelease to contain tag number, fix typo in readme name

+++ mkrelease.sh
@@ -3,7 +3,14 @@
 # After building artifacts using Visual Studio, build a zip file
 # in Artifacts folder, for associating with a release tag.
 
-ZIP=PianoRes-win-x64-vst3.zip
+if [ $# -lt 1 ] ; then
+    echo "expecting version number, with v prefix"
+    exit 1
+fi
+
+VERSION=$1
+
+ZIP=PianoRes-$VERSION-win-x64-vst3.zip
 set -ex
 mkdir -p Artifacts
 rm -f Artifacts/$ZIP
@@ -14,4 +21,4 @@ zip -r $OLDPWD/Artifacts/$ZIP PianoRes.vst3
 cd -
 
 cd release-readme
-zip -r $OLDPWD/Artifacts/$ZIP README-win-x64-fst3.txt
+zip -r $OLDPWD/Artifacts/$ZIP README-win-x64-vst3.txt
